requirements.txt
Django>=4.2
psycopg2-binary
djangorestframework-simplejwt
djangorestframework